Extracted Fitting. Moved a parametrization method into CurveHelpers.
